pyproject.toml
[tool.poetry]
name = "aicode"
version = "0.1.0"
description = ""
authors = ["wangyuxin <wangyuxin@mokahr.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.3"
datasets = "^2.3.2"
transformers = "^4.20.1"
PyYAML = "^6.0"
tensorboard = "^2.9.1"
torchmetrics = "^0.9.2"
rich = "^12.4.4"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.0"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
include = '\.pyi?$'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
